задание 2-3: фрагмент для вывода заметок, встраиваем в основную активити. "массив" зметок и его заполнение пришлось перенести в этот фрагмент
